feat(token): transfer token to a standalone module
