+++ src/modules/auth/index.ts
@@ -1,4 +1,3 @@
-export * from './dto';
 export * from './interface';
 export { default as AuthRouter } from './auth.router';
 export * from './auth.type';
