+++ src/modules/auth/interface/index.ts
@@ -1,4 +1,1 @@
 export * from './auth-service.interface';
-export * from './refresh-token-repository.interface';
-export * from './refresh-token.interface';
-export * from './token-service.interface';
