--- src/modules/auth/service/index.ts
@@ -1,2 +1,0 @@
-export { default as AuthService } from './auth.service';
-export { default as TokenService } from './token.service';
