+++ src/modules/index.di.ts
@@ -1,4 +1,5 @@
 import './auth/auth.di';
 import './notification/notification.di';
 import './platform/platform.di';
+import './token/token.di';
 import './user/user.di';
